+++ lib/class/ApiMethods/AlbumsMethod.php
@@ -67,14 +67,14 @@ final class AlbumsMethod
 
         $albums  = Api::$browse->get_objects();
         if (empty($albums)) {
-            Api::error(T_('No Results'), '4704', self::ACTION, 'empty', $input['api_format']);
+            Api::empty('album', $input['api_format']);
 
             return false;
         }
-        $user    = User::get_from_username(Session::username($input['auth']));
-        $include = (is_array($input['include'])) ? $input['include'] : explode(',', (string) $input['include']);
 
         ob_end_clean();
+        $user    = User::get_from_username(Session::username($input['auth']));
+        $include = (is_array($input['include'])) ? $input['include'] : explode(',', (string) $input['include']);
         switch ($input['api_format']) {
             case 'json':
                 JSON_Data::set_offset($input['offset']);