+++ lib/class/ApiMethods/FollowingMethod.php
@@ -72,18 +72,19 @@ final class FollowingMethod
         }
 
         $users = $user->get_following();
-        if (!count($users)) {
-            Api::error(T_('No Results'), '4704', self::ACTION, 'empty', $input['api_format']);
-        } else {
-            debug_event(self::class, 'User is following:  ' . print_r($users), 1);
-            ob_end_clean();
-            switch ($input['api_format']) {
-                case 'json':
-                    echo JSON_Data::users($users);
-                    break;
-                default:
-                    echo XML_Data::users($users);
-            }
+        if (empty($users)) {
+            Api::empty('user', $input['api_format']);
+
+            return false;
+        }
+
+        ob_end_clean();
+        switch ($input['api_format']) {
+            case 'json':
+                echo JSON_Data::users($users);
+                break;
+            default:
+                echo XML_Data::users($users);
         }
         Session::extend($input['auth']);