+++ lib/class/ApiMethods/GetSimilarMethod.php
@@ -79,7 +79,7 @@ final class GetSimilarMethod
             $objects[] = $child['id'];
         }
         if (empty($objects)) {
-            Api::error(T_('No Results'), '4704', self::ACTION, 'empty', $input['api_format']);
+            Api::empty($type, $input['api_format']);
 
             return false;
         }