+++ lib/class/catalog.class.php
@@ -1789,10 +1789,12 @@ abstract class Catalog extends database_object
             case 'album':
                 $tags = self::getSongTags('album', $libitem->id);
                 Tag::update_tag_list(implode(',', $tags), 'album', $libitem->id, false);
+                $libitem->update_time();
                 break;
             case 'artist':
                 $tags = self::getSongTags('artist', $libitem->id);
                 Tag::update_tag_list(implode(',', $tags), 'artist', $libitem->id, false);
+                $libitem->update_time();
                 break;
         } // end switch type
 
@@ -2950,6 +2952,27 @@ abstract class Catalog extends database_object
                     }
                 }
                 break;
+            case 'update_all_file_tags':
+                $catalogs = self::get_catalogs();
+                // Intentional break fall-through
+            case 'update_file_tags':
+                $write_id3 = AmpConfig::get('write_id3', false);
+                AmpConfig::set('write_id3', 'true', true);
+                $write_id3_art = AmpConfig::get('write_id3_art', false);
+                AmpConfig::set('write_id3_art', 'true', true);
+                foreach ($catalogs as $catalog_id) {
+                    $catalog = self::create_from_id($catalog_id);
+                    if ($catalog !== null) {
+                        $song_ids = $catalog->get_song_ids();
+                        foreach ($song_ids as $song_id) {
+                            $song = new Song($song_id);
+                            $song->format();
+                            $song->write_id3();
+                        }
+                    }
+                }
+                AmpConfig::set('write_id3', $write_id3, true);
+                AmpConfig::set('write_id3', $write_id3_art, true);
         }
 
         // Remove any orphaned artists/albums/etc.