+++ lib/class/stats.class.php
@@ -264,17 +264,10 @@ class Stats
         $sqlres = array($user_id);
 
         $sql = "SELECT `object_count`.`id`, `object_count`.`object_type`, `object_count`.`object_id`, " .
-               "`object_count`.`user`, `object_count`.`agent`, `object_count`.`date`, `song`.`time`, " .
+               "`object_count`.`user`, `object_count`.`agent`, `object_count`.`date`, " .
                "`object_count`.`count_type` FROM `object_count` " .
-               "LEFT JOIN `song` ON `song`.`id` = `object_count`.`object_id` ";
-        if (AmpConfig::get('catalog_disable')) {
-            $sql .= "LEFT JOIN `catalog` ON `catalog`.`id` = `song`.`catalog` ";
-        }
-        $sql .= "WHERE `object_count`.`user` = ? AND `object_count`.`object_type` " .
-                "IN ('song', 'video', 'podcast_episode') AND `object_count`.`count_type` IN ('stream', 'skip') ";
-        if (AmpConfig::get('catalog_disable')) {
-            $sql .= "AND `catalog`.`enabled` = '1' ";
-        }
+               "WHERE `object_count`.`user` = ? AND `object_count`.`object_type` " .
+               "IN ('song', 'video', 'podcast_episode') AND `object_count`.`count_type` IN ('stream', 'skip') ";
         if ($agent) {
             $sql .= "AND `object_count`.`agent` = ? ";
             array_push($sqlres, $agent);
@@ -288,6 +281,24 @@ class Stats
 
         return Dba::fetch_assoc($db_results);
     } // get_last_play
+
+    /**
+     * get_time
+     *
+     * get the time for the object (song, video, podcast_episode)
+     * @param integer $object_id
+     * @param string $object_type
+     * @return integer
+     */
+    public static function get_time($object_id, $object_type)
+    {
+        $sql = "SELECT `time` FROM `$object_type` " .
+               "WHERE `id` = ?";
+        $db_results = Dba::read($sql, array($object_id));
+        $results    = Dba::fetch_assoc($db_results);
+
+        return (int) $results['time'];
+    } // get_time
 
     /**
      * skip_last_play
@@ -328,12 +339,13 @@ class Stats
             return false;
         }
         $previous  = self::get_last_play($user, $agent, $date);
+        $last_time = self::get_time($previous['object_id'], $previous['object_type']);
         $diff      = $date - (int) $previous['date'];
         $item_time = $object->time;
-        $skip_time = AmpConfig::get_skip_timer($previous['time']);
+        $skip_time = AmpConfig::get_skip_timer($last_time);
 
         // if your last song is 30 seconds and your skip timer if 40 you don't want to keep skipping it.
-        if ($previous['time'] > 0 && $previous['time'] < $skip_time) {
+        if ($last_time > 0 && $last_time < $skip_time) {
             return true;
         }
 
@@ -345,7 +357,7 @@ class Stats
         }
 
         // when the difference between recordings is too short, the previous object has been skipped, so note that
-        if (($diff < $skip_time || ($diff < $skip_time && $previous['time'] > $skip_time))) {
+        if (($diff < $skip_time || ($diff < $skip_time && $last_time > $skip_time))) {
             debug_event('stats.class', 'Last ' . $previous['object_type'] . ' played within skip limit (' . $diff . '/' . $skip_time . 's). Skipping {' . $previous['object_id'] . '}', 3);
             self::skip_last_play($previous['date'], $previous['agent'], $previous['user']);
             // delete song, podcast_episode and video from user_activity to keep stats in line
@@ -628,6 +640,7 @@ class Stats
             case 'tvshow_episode':
             case 'movie':
             case 'playlist':
+            case 'podcast_episode':
                 return $type;
             default:
                 return 'song';