+++ templates/show_recently_played.inc.php
@@ -119,7 +119,6 @@ foreach ($data as $row) {
             </a>
         </td>
         <td class="cel_lastplayed"><?php echo $time_string; ?></td>
-        </td>
     </tr>
 <?php
     ++$nb;